project/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.2.12
djangorestframework==3.13.1
geolocation-python==0.2.2
googlemaps==4.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
numpy==1.19.5
pandas==1.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-google-places==1.4.2
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
websockets==9.1
zope.interface==5.4.0
